lambda/requirements.txt
Pillow==9.1.1
boto3==1.24.0
